+++ plugins/Tour/tests/Integration/ChallengeTest.php
@@ -66,32 +66,36 @@ class ChallengeTest extends IntegrationTestCase
 
     public function test_skip()
     {
-        $this->assertFalse($this->challenge->isSkipped());
-        $this->assertFalse($this->challenge2->isSkipped());
-        $this->assertFalse($this->challenge->isCompleted());
-        $this->assertFalse($this->challenge2->isCompleted());
+        $login = 'foo';
+        $this->assertFalse($this->challenge->isSkipped($login));
+        $this->assertFalse($this->challenge2->isSkipped($login));
+        $this->assertFalse($this->challenge->isCompleted($login));
+        $this->assertFalse($this->challenge2->isCompleted($login));
 
-        $this->challenge2->skipChallenge();
+        $this->challenge2->skipChallenge($login);
 
-        $this->assertFalse($this->challenge->isSkipped());
-        $this->assertTrue($this->challenge2->isSkipped());
-        $this->assertFalse($this->challenge->isCompleted());
-        $this->assertFalse($this->challenge2->isCompleted());
+        $this->assertFalse($this->challenge->isSkipped($login));
+        $this->assertTrue($this->challenge2->isSkipped($login));
+        $this->assertFalse($this->challenge2->isSkipped('barbaz'));
+        $this->assertFalse($this->challenge->isCompleted($login));
+        $this->assertFalse($this->challenge2->isCompleted($login));
     }
 
     public function test_complete()
     {
-        $this->assertFalse($this->challenge->isSkipped());
-        $this->assertFalse($this->challenge2->isSkipped());
-        $this->assertFalse($this->challenge->isCompleted());
-        $this->assertFalse($this->challenge2->isCompleted());
+        $login = 'foo';
+        $this->assertFalse($this->challenge->isSkipped($login));
+        $this->assertFalse($this->challenge2->isSkipped($login));
+        $this->assertFalse($this->challenge->isCompleted($login));
+        $this->assertFalse($this->challenge2->isCompleted($login));
 
-        $this->challenge->setCompleted();
+        $this->challenge->setCompleted($login);
 
-        $this->assertFalse($this->challenge->isSkipped());
-        $this->assertFalse($this->challenge2->isSkipped());
-        $this->assertTrue($this->challenge->isCompleted());
-        $this->assertFalse($this->challenge2->isCompleted());
+        $this->assertFalse($this->challenge->isSkipped($login));
+        $this->assertFalse($this->challenge2->isSkipped($login));
+        $this->assertTrue($this->challenge->isCompleted($login));
+        $this->assertFalse($this->challenge->isCompleted('barbaz'));
+        $this->assertFalse($this->challenge2->isCompleted($login));
     }
 
 }